Don't remake cached variables for package use
Included files for Makefile cause some trouble if user uses a package of
mduemized project, because make tries to remake cached variables but its
dependencies don't exist in such situation.

Fix this problem by removing the rule to generate cached variables if
the current directory is not controlled in git.