Fix value format of $(test_cases)
In the old version, $(test_cases) contains a list of values which are
like "test/{case}".  But it is not clear for further processes.  Fix to
use "{case}" instead.

Also there is a bug in the old version.  If there is a file with suffix
".expected" which is not in the "test" directory, it is treated as
a test case, but it's wrong.  Fix to select files in the "test"
directory to compupte $(test_cases) explicitly.