test: pack: Fix to normalize both outputs
There are two problems which cause unwanted difference of outputs:

(a) "unzip -Z1" outputs lines which contain only the name of
a directory, while "git ls-files" doesn't so.  Fix to remove such lines.

(b) The order of output from "git ls-files" is not the same as the one
from "unzip -Z1" if the name of a file starts with lowercase and the
name of another file starts with uppercase.  Fix to sort both output
with sort(1) explicitly.