examples: Override configuration for vim-vspec
DEPS_{project}_URI should be written in absolute URI, but
DEP_vim_vspec_URI is written in relative URI for ease of development.
fetch-deps parses a relative URI like git submodule, but it doesn't work
for package use if the current directory is examples/vim-vspec/.

To avoid confusion, override configuration for vim-vspec in the example.