Fix TEST_RULE_DEPS to be cumulative
TEST_RULE_DEPS can be overridden by users to customize dependencies.
But if it is overridden, the default value of TEST_RULE_DEPS is
completely lost.  It is not desired effect; the default value should
also be used as well as user-specified value.

Fix this problem by changing the role of TEST_RULE_DEPS.
Use TEST_RULE_DEPS for user-specified dependencies,
add builtin_test_rule_deps for the default dependencies, and
add all_test_rule_deps for rule definitions.