Fix a bug in the default target definition
In old versions, there is an assumption that no rule is not defined
before including Makefile.mduem.  But it's not useful for practical use
and users often forget about the assumption.

Fix to use .DEFAULT_GOAL to define the default target.  It also works
even if some rule is defined before including Makefile.mduem.