add a prefix mapping option to RegEdit.
This lets us define a mapping to <Plug>(RegEditPrefix) instead
of <Plug>(RegEditPostfix) so that we can use register selection
conventions similar to other normal-mode commands.  i.e. "a<leader>E
to edit the a register as opposed to <leader>Ea.

Both are fairly regularly used in vim (i.e. qa vs "ayy) so let the
user decide which they want while retaining the postfix-style
as the default.